added sql2 and func2 to original files, deleted extra files
