+++ app/misc_func.h
@@ -34,6 +34,19 @@ void create_account();              //create account for user and save to databa
 bool valid_ssn(std::string ssn);
 
 
+
+void loginas_IQC_inspector();
+void loginas_OQC_inspector();
+//void loginas_technician();
+void loginas_QAdirector();
+void loginas_inv_associate();
+
+void display_IQCinsp_menu();
+void display_OQCinsp_menu();
+void display_tech_menu();
+void display_QAdir_menu();
+void display_invassoc_menu();
+
 /*
 functions to show menu for each entity type
 _view()
