--- app/misc_func2.h
@@ -1,21 +1,0 @@
-#ifndef MISC_FUNC2_H
-#define MISC_FUNC2_H
-
-#include <iostream>
-#include <fstream>
-#include "sql_func.h"
-
-void loginas_IQC_inspector();
-void loginas_OQC_inspector();
-//void loginas_technician();
-void loginas_QAdirector();
-void loginas_inv_associate();
-
-void display_IQCinsp_menu();
-void display_OQCinsp_menu();
-void display_tech_menu();
-void display_QAdir_menu();
-void display_invassoc_menu();
-  
-
-#endif //MISC_FUNC2_H
