+++ app/sql_func.h
@@ -20,6 +20,14 @@ bool ssn_exists(int ssn);               //return boolean indicating if ssn exist
 bool verify_user(std::string user, std::string pw);         //return boolean indicating if username and password input exist in a tuple
 void create_user(int ssn, std::string name, std::string user, std::string pw);  //add tuple to db using user inputs
 
+void view_IQC_list();
+void create_inspection(int pn, std::string pdesc, int insp_area, std::string requirements, std::string result_type, int sample_size);
+void approve_inspection(int insp_num);
+void view_inspection(int insp_num);
+void move_to_IPQC(int pn);
 
+void receive_material(int pn);
+void backflush_product(int pn);
+void move_to_IQC(int pn);
 
 #endif
